teradata/requirements.txt
apache-beam[gcp]
google-cloud-secret-manager